refactor(observability): split the connection recorders out of migrations
renderer-diagnostic-migrations.ts went 121 -> 405 lines in this PR, which
is how a shared wrapper module becomes a junk drawer. The connection and
stream lifecycle half moves to renderer-diagnostics-connection.ts: the
session SSE stream, the runtime bootstrap handshake, and the terminal and
feed websockets. That is exactly this PR's first commit, so the seam is
the one the work already had. 405 -> 253, new file 164.

Five call sites are repointed at the new module rather than served a
re-export, so each recorder has one import path.

Two files keep their instrumentation and take an allowlist entry instead,
because neither has a seam worth inventing:

  session-stream-flush-apply.ts 401 -> 420  (already ledgered; +19 for the
      batch-applied and gap-detected records, count updated in place)
  terminal-stream-registry.ts   395 -> 417  (new entry; +22 woven into
      three existing callbacks)
