+++ scripts/frontend_structure_allowlist.txt
@@ -39,3 +39,4 @@ apps/packages/product-client/src/components/playground/GitReviewV2Playground.tsx
 apps/packages/design/src/tokens.ts 2143 single token authority: one entry per token with provenance by design; shape is a flat manifest, not logic
 apps/packages/product-client/src/hooks/agents/workflows/use-harness-auth-editor.ts 410 harness auth editor grew with v2 single-PUT replace semantics + provider-hint rows; split pending
 apps/packages/product-client/src/primitives/icons/core.tsx 483 existing consolidated icon module pending split
+apps/packages/product-client/src/lib/infra/terminals/terminal-stream-registry.ts 417 terminal/feed stream registry (+open/close/error records woven into the existing callbacks; no seam short of splitting the registry itself)
